src/services/industry_context_service/requirements.txt
fastapi==0.95.1
uvicorn[standard]==0.22.0
pymongo==4.4.0
motor==3.1.1
boto3==1.26.152
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
python-pptx==0.7.9
sentence-transformers==2.2.2
chromadb==0.4.1
pydantic==2.1.2